# Faster conversions from string to decimal,int,long