--- a/Common/Data/Market/Tick.cs
+++ b/Common/Data/Market/Tick.cs
@@ -184,13 +184,13 @@ namespace QuantConnect.Data.Market
             var csv = line.Split(',');
             DataType = MarketDataType.Tick;
             Symbol = symbol;
-            Time = baseDate.Date.AddMilliseconds(Convert.ToInt32(csv[0]));
+            Time = baseDate.Date.AddMilliseconds(csv[0].ToInt32());
             Value = csv[1].ToDecimal()/10000m;
             TickType = TickType.Trade;
-            Quantity = Convert.ToInt32(csv[2]);
+            Quantity = csv[2].ToInt32();
             Exchange = csv[3].Trim();
             SaleCondition = csv[4];
-            Suspicious = Convert.ToInt32(csv[5]) == 1;
+            Suspicious = csv[5].ToInt32() == 1;
         }
 
 
@@ -212,10 +212,10 @@ namespace QuantConnect.Data.Market
                 { 
                     case SecurityType.Equity:
                         Symbol = config.Symbol;
-                        Time = date.Date.AddMilliseconds(Convert.ToInt64(csv[0]));
+                        Time = date.Date.AddMilliseconds(csv[0].ToInt64());
                         Value = config.GetNormalizedPrice(csv[1].ToDecimal() / 10000m);
                         TickType = TickType.Trade;
-                        Quantity = Convert.ToInt32(csv[2]);
+                        Quantity = csv[2].ToInt32();
                         if (csv.Length > 3)
                         {
                             Exchange = csv[3];
@@ -227,7 +227,7 @@ namespace QuantConnect.Data.Market
                     case SecurityType.Forex:
                         Symbol = config.Symbol;
                         TickType = TickType.Quote;
-                        Time = date.Date.AddMilliseconds(Convert.ToInt64(csv[0]));
+                        Time = date.Date.AddMilliseconds(csv[0].ToInt64());
                         BidPrice = csv[1].ToDecimal();
                         AskPrice = csv[2].ToDecimal();
                         Value = (BidPrice + AskPrice) / 2;