Iterations: 10,000,000

Convert.ToInt32(): 1445.00 msec
New str.ToInt32(): 69.00 msec

Convert.ToInt64(): 1423.00 msec
New str.ToInt64(): 80.00 msec

Convert.ToDecimal(): 2608.00 msec
Old str.ToDecimal(): 2392.00 msec
New str.ToDecimal(): 189.00 msec